+++ NEWS.md
@@ -1,6 +1,6 @@
 # random.cdisc.data 0.3.6.9000
 
-* issues with no news:
+* ADTTE was updated. There is a new parameter in this data set: "PARAM == Duration of Confirmed Response".
 
 # random.cdisc.data 0.3.6